feat: restructure HomePage and integrate HomeHeader and AppBottomNavigationBar
