+++ app/ui/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'presentation/pages/home/home_page.dart';
+import 'presentation/pages/home_page/home_page.dart';
 
 void main() {
   runApp(const MainApp());
