--- app/ui/lib/presentation/pages/home/home_page.dart
@@ -1,190 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Column(
-          children: [
-            headerSection(),
-            productList(),
-          ],
-        ),
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.home),
-            label: 'Home',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.category),
-            label: 'Categories',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.shopping_cart),
-            label: 'Cart',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.person),
-            label: 'Profile',
-          ),
-        ],
-        currentIndex: 0, // Set the current index as needed
-        onTap: (index) {
-          // Handle navigation on tap
-        },
-        backgroundColor: Colors.white, // Set the background color
-        selectedItemColor: Colors.blue, // Set the color of the selected item
-        unselectedItemColor: Colors.grey, // Set the color of the unselected items
-      ),
-    );
-  }
-
-  Widget headerSection() {
-    // Contain only "Somba" text and a search bar input
-    return Column(
-      children: [
-        Container(
-          padding: const EdgeInsets.all(10),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Expanded(
-                child: Container(
-                  height: 40,
-                  decoration: BoxDecoration(
-                    color: Colors.grey[200],
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                  child: const TextField(
-                    decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText: 'Search on Somba.com',
-                      prefixIcon: Icon(Icons.search),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        // Category list
-        SizedBox(
-          height: 50,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            children: [
-              categoryItem('All'),
-              categoryItem('Fashion'),
-              categoryItem('Electronics'),
-              categoryItem('Home'),
-              categoryItem('Beauty'),
-              categoryItem('Toys'),
-              categoryItem('Sports'),
-              categoryItem('Books'),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget categoryItem(String title) {
-    return Container(
-      margin: const EdgeInsets.only(left: 10),
-      padding: const EdgeInsets.symmetric(horizontal: 10),
-      decoration: BoxDecoration(
-        color: Colors.grey[200],
-        borderRadius: BorderRadius.circular(10),
-      ),
-      child: Center(
-        child: Text(title),
-      ),
-    );
-  }
-
-  // Product list
-  Widget productList() {
-    return Expanded(
-      child: GridView.builder(
-        gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-          crossAxisCount: 2,
-          childAspectRatio: 0.7,
-        ),
-        itemCount: 10,
-        itemBuilder: (context, index) {
-          return productItem();
-        },
-      ),
-    );
-  }
-
-  Widget productItem() {
-    return GestureDetector(
-      onTap: () {
-        // Navigate to product details
-      },
-      child: Container(
-        margin: const EdgeInsets.all(10),
-        padding: const EdgeInsets.all(10),
-        decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.circular(10),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.grey.withOpacity(0.5),
-              spreadRadius: 2,
-              blurRadius: 5,
-              offset: const Offset(0, 3),
-            ),
-          ],
-        ),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Flexible(
-              child: Image.network(
-                'https://via.placeholder.com/150',
-                width: double.infinity,
-                fit: BoxFit.cover,
-              ),
-            ),
-            const SizedBox(height: 10),
-            const Text(
-              'Product Name',
-              style: TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-              ),
-              overflow: TextOverflow.ellipsis,
-            ),
-            const SizedBox(height: 5),
-            const Text(
-              'Price: \$100',
-              style: TextStyle(
-                fontSize: 14,
-                color: Colors.green,
-              ),
-              overflow: TextOverflow.ellipsis,
-            ),
-            const SizedBox(height: 5),
-            SizedBox(
-              width: double.infinity,
-              child: ElevatedButton(
-                onPressed: () {
-                  // Add to cart action
-                },
-                child: const Text('Add to Cart'),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
